#24 Add a --rf=<goal> option to allow start execution from specific goal
This --rf option takes as value the goal name from which execution should restart.
When there is a failure, it doesn't print yet the command to execute to resume from this point

--rf option is currently stateless which means that it doesn't support (yet?) --rf alone.
This means that you can't launch the build with only --rf option (no goal information provided) to restart from last failing goal.